requirements.txt
asgiref==3.8.1
certifi==2022.12.7
cffi
charset-normalizer==3.0.1
crispy-bootstrap5==2024.10
cryptography==39.0.1
defusedxml==0.7.1
Django==5.1.2
django-allauth==0.58.2
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-storages==1.14.2
boto3==1.34.1
gunicorn==23.0.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sentry-sdk==1.39.1
django-prometheus==2.3.1
sqlparse==0.4.3
typing_extensions==4.9.0
urllib3==1.26.14
whitenoise==6.7.0
python-json-logger==2.0.7

# API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
django-cors-headers==4.3.1
django-filter==23.5
drf-nested-routers==0.93.5
drf-yasg==1.21.7

# Authentication
dj-rest-auth==5.0.2
